requirements.txt
python-dotenv
streamlit
langchain
langchain_openai
langgraph
langchain_community
azure-core
azure-search-documents
azure-ai-textanalytics
azure-storage-blob
python-docx
markdown2
matplotlib
bs4
openai
requests
pandas
openpyxl
urllib3
plotly
seaborn
chardet